+++ BrotBoxEngine/GeneralPurposeAllocatorTest.h
@@ -67,7 +67,7 @@ namespace bbe {
 					{
 						if (list.getLength() > 0)
 						{
-							size_t index = rand.randomInt(list.getLength());
+							size_t index = (size_t)rand.randomInt((int)list.getLength());
 							gpa.deallocateObjects(list[index]);
 							list.removeIndex(index);
 						}
